scripts/vonds-ads-api/requirements.txt
google-ads>=24.0.0,<25.0.0
google-auth-oauthlib>=1.2.0
PyYAML>=6.0
